refactor section trains, alignment
put examples at the end because the exemplary code should be so close
together that it is easy to compare. this was not the case anymore when
the examples were in between the paragraphs; it annoyed me.

undo swapping the nvv-fork and capped-fork sections as the examples do
not require it anymore and just fixing an argument is not as big a
change as making forks behave monadically instead of dyadically.